:alembic: ci: Use Travis CI to build and deploy Hugo docs (closes #8)
This commit sets up a Travis CI config that installs Hugo into the build
environment, builds our docs, and if the test runs on the `master`
branch, it will push the built assets to a `gh-pages` branch on the
librecorps/resources repo. Once this commit is merged, we can switch
over the build settings to use that branch to serve our docs.

Closes #8.

Signed-off-by: Justin W. Flory <git@jwf.io>