:boom: Remove contact page
We don't want people to ever contact us. Just kidding. But to get this
working, it requires using a proprietary service. I'd rather strip it
out and not have someone stumble on a broken contact form and expect it
go somewhere. I removed it from the theme and all references in our site
config.

Signed-off-by: Justin W. Flory <git@jwf.io>