:name_badge: Rename product-management to project-management
It is a semantic nitpick, but since we are communicating "project
management" resources to our current LibreCorps cohort, I want to be
consistent in the language we use in our other resources like the
milestone roadmap and rubric. I also think it is a little more "open
source"-y to call it project instead of product. While products can be
open source, anything done in the open can be considered a project, even
if it isn't related to the bottom line of your organization.

It is totally a semantic nitpick, but again, I just want to be
consistent in our vocabulary. :slightly_smiling_face:

Signed-off-by: Justin W. Flory <git@jwf.io>