:gear: WIP: Add Open Hardware research draft
This commit adds a brain-dump of all the content collected from our
network of Open Hardware folks who have given us advice and pointers on
how to navigate the Open Hardware space. Currently it is massively
disorganized and mostly just a list of "things".

One of the things I now love about Hugo is I marked this page as a
"draft". So, I can get this stuff out of my head and into `git` without
it being published in our docs site. If you run `hugo server -D`
locally, you can generate a version of the site with this page included.
But in the production environment on GitHub Pages, this page will not
exist. Hooray! :tada:

I really needed to get this down somewhere because I keep putting it off
since it is a lot of work. But this will make it easier for me to
iterate on something and hopefully turn it into something more
meaningful in the next few days.

Signed-off-by: Justin W. Flory <git@jwf.io>