:foggy: :boom: Retire Legal section
This content now lives, with edits and improvements, here:

    https://jwf.io/heuristics/foss/

    https://github.com/jwflory/heuristics/commit/d6e2e42f04cbcb61e9af49d290d999cb341f6e2d

It might look weird to drop this content from LibreCorps and move to a
personal repo. However, I am using a Free Culture license (CC BY-SA) on
the above repo. So the information remains Free. But lately, I am
maintaining this information in multiple places. I am flipping between
three different repos depending on what I want to talk about.

So, this commit is part of a gradual effort I am undertaking to simplify
my life by moving this content to a personal repo, since I am the
primary maintainer of this information at the moment. The bus factor of
LibreCorps feels a little high on me at the moment, and I want this
information / work to live beyond in the future, for whatever purposes I
might wish to use it for in the future.

This is a long explanation for a commit message, but I am trying to add
important context for why I am moving this content from the LibreCorps
collection of info into my own personal repo.

Signed-off-by: Justin W. Flory <git@jwf.io>