:beetle: ci: Upgrade Go and Hugo to latest versions
The Travis CI pipeline ends up timing out after ten minutes when it
tries to build the site. Why? I couldn't tell you. But this is one of
the reasons I am keen to move everything off of Travis CI. This is
annoying.

This commit is an attempt to get the CI pipeline working by updating to
a newer Golang version and updating the Hugo binary to something
released last week.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>